exercise: Replaced apply_diffs function get_diffs function
