+++ exercise/src/cmd/generate.rs
@@ -85,136 +85,6 @@ fn update_cargo_toml(exercise_name: &str, exercise_path: &Path, canonical_data: 
         .expect("Failed to update Cargo.toml file");
 }
 
-fn generate_property_body<'a>(
-    property_functions: &mut HashMap<&'a str, String>,
-    case: &'a JsonValue,
-) {
-    if let Some(property) = case.get("property") {
-        let property = property.as_str().unwrap();
-
-        if property_functions.contains_key(property) {
-            return;
-        }
-
-        let property_function_body = format!(
-            "\
-             /// Process a single test case for the property `{property}`\n\
-             ///\n\
-             /// All cases for the `{property}` property are implemented\n\
-             /// in terms of this function.\n\
-             /// \n\
-             /// Note that you'll need to both name the expected transform which\n\
-             /// the student needs to write, and name the types of the inputs and outputs.\n\
-             /// While rustc _may_ be able to handle things properly given a working example,\n\
-             /// students will face confusing errors if the `I` and `O` types are not concrete.\n\
-             /// \n\
-             fn process_{property_formatted}_case<I, O>(input: I, expected: O) {{\n\
-             //  typical implementation:\n\
-             //  assert_eq!(\n\
-             //      student_{property_formatted}_func(input),\n\
-             //      expected\n\
-             //  )\n    unimplemented!()\n\
-             }}\n\
-             \n\
-             ",
-            property = property,
-            property_formatted = utils::format_exercise_property(property),
-        );
-
-        property_functions.insert(property, property_function_body);
-    }
-}
-
-// Depending on the type of the item variable,
-// transform item into corresponding Rust literal
-fn into_literal(item: &JsonValue, use_maplit: bool) -> String {
-    if item.is_string() {
-        format!("\"{}\"", item.as_str().unwrap())
-    } else if item.is_array() {
-        format!(
-            "vec![{}]",
-            item.as_array()
-                .unwrap()
-                .iter()
-                .map(|item| into_literal(item, use_maplit))
-                .collect::<Vec<String>>()
-                .join(", ")
-        )
-    } else if item.is_number() || item.is_boolean() || item.is_null() {
-        format!("{}", item)
-    } else if !use_maplit {
-        let key_values = item
-            .as_object()
-            .unwrap()
-            .iter()
-            .map(|(key, value)| {
-                format!(
-                    "hm.insert(\"{}\", {});",
-                    key,
-                    into_literal(value, use_maplit)
-                )
-            }).collect::<String>();
-
-        format!(
-            "{{let mut hm = ::std::collections::HashMap::new(); {} hm}}",
-            key_values
-        )
-    } else {
-        let key_values = item
-            .as_object()
-            .unwrap()
-            .iter()
-            .map(|(key, value)| format!("\"{}\"=>{}", key, into_literal(value, use_maplit)))
-            .collect::<Vec<String>>()
-            .join(",");
-
-        format!("hashmap!{{{}}}", key_values)
-    }
-}
-
-fn generate_test_function(case: &JsonValue, use_maplit: bool) -> String {
-    let description = case.get("description").unwrap().as_str().unwrap();
-
-    let property = case.get("property").unwrap().as_str().unwrap();
-
-    let comments = if let Some(comments) = case.get("comments") {
-        if comments.is_array() {
-            comments
-                .as_array()
-                .unwrap()
-                .iter()
-                .map(|line| format!("/// {}", line))
-                .collect::<String>()
-        } else {
-            format!("/// {}\n", comments.as_str().unwrap())
-        }
-    } else {
-        "".to_string()
-    };
-
-    let input = into_literal(case.get("input").unwrap(), use_maplit);
-
-    let expected = into_literal(case.get("expected").unwrap(), use_maplit);
-
-    format!(
-        "#[test]\n\
-         #[ignore]\n\
-         /// {description}\n\
-         {comments}
-         fn test_{description_formatted}() {{\n\
-         process_{property}_case({input}, {expected});\n\
-         }}\n\
-         \n\
-         ",
-        description = description,
-        description_formatted = utils::format_exercise_description(description),
-        property = utils::format_exercise_property(property),
-        comments = comments,
-        input = input,
-        expected = expected
-    )
-}
-
 // Generate test suite using the canonical data
 fn generate_tests_from_canonical_data(
     exercise_name: &str,
@@ -258,14 +128,27 @@ fn generate_tests_from_canonical_data(
     for case in cases.as_array().unwrap().iter() {
         if let Some(sub_cases) = case.get("cases") {
             for sub_case in sub_cases.as_array().unwrap().iter() {
-                generate_property_body(&mut property_functions, &sub_case);
+                if let Some(property) = sub_case.get("property") {
+                    let property = property.as_str().unwrap();
 
-                test_functions.push(generate_test_function(&sub_case, use_maplit));
+                    if !property_functions.contains_key(property) {
+                        property_functions
+                            .insert(property, utils::generate_property_body(property));
+                    }
+                }
+
+                test_functions.push(utils::generate_test_function(&sub_case, use_maplit));
             }
         } else {
-            generate_property_body(&mut property_functions, &case);
+            if let Some(property) = case.get("property") {
+                let property = property.as_str().unwrap();
 
-            test_functions.push(generate_test_function(&case, use_maplit));
+                if !property_functions.contains_key(property) {
+                    property_functions.insert(property, utils::generate_property_body(property));
+                }
+            }
+
+            test_functions.push(utils::generate_test_function(&case, use_maplit));
         }
     }
